spikes/002-rustfs-s3-gateway-modal/requirements.txt
modal
boto3
botocore
python-dotenv
pytest
